[W5.3b][W14-A1]TEO WEN ZONG
Create Contact.java to serve as a parent class for Address.java, Name.java, Phone.java
